sync

+++ DataCrawling/file_utils.py
@@ -22,14 +22,21 @@ def read_data_from_json_file(file_name):
 
 # Đọc data từ file html
 def read_data_from_html_file(file_name):
-    with open(file_name, encoding='utf-8') as html_file:
-        data = html_file.read()
+    data = ""
+    if not os.path.exists(file_name):
+        print(f"FAIL: read_data_from_html_file: File path is not exist {file_name}")
+    else:
+        with open(file_name, encoding='utf-8') as html_file:
+            data = html_file.read()
     return data
 
 # Ghi data vào file json
 def write_data_to_json_file(data, file_name):
-    with open(file_name, 'w', encoding='utf-8') as json_file:
-        json.dump(data, json_file, ensure_ascii=False, indent=4)
+    if not os.path.exists(file_name):
+        print(f"FAIL: read_data_from_html_file: File path is not exist {file_name}")
+    else:
+        with open(file_name, 'w', encoding='utf-8') as json_file:
+            json.dump(data, json_file, ensure_ascii=False, indent=4)
 
 # Ghi data vào file html
 def write_data_to_html_file(data, file_name):
@@ -46,6 +53,9 @@ def append_data_to_json_file(data, file_name):
         json_file.truncate()
 
 def song_link_to_relative_html_file_path(link):
+    '''
+    Chuyển đường dẫn đến bài hát từ dạng link sang dạng đường dẫn tương đối đến file html: <artist_letter>/<artist>/<song>.html
+    '''
     artist_letter = link.split("/")[-2][0]
     artist = link.split("/")[-2]
     song = link.split("/")[-1]
